# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/categories": {
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryBodyDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "create a new wallet category",
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "operationId": "CategoryController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "list all wallet categories",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/categories/{slug}": {
      "put": {
        "operationId": "CategoryController_edit",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCategoryBodyDTO"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "edit an wallet category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/transactions": {
      "post": {
        "operationId": "CreateTransactionController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionBodyDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "create a new transaction",
        "tags": [
          "Transactions"
        ]
      },
      "get": {
        "operationId": "ListWalletTransactionsController_handle",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "list wallet transactions with pagination and optional filters",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}": {
      "put": {
        "operationId": "EditWalletTransactionController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTransactionBodyDTO"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "edit a transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Finance API",
    "description": "Finance manager API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCategoryBodyDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "EditCategoryBodyDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateTransactionBodyDTO": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "x-nestjs_zod-has-const": true
          },
          "title": {
            "type": "string",
            "x-nestjs_zod-has-const": true
          },
          "description": {
            "type": "string",
            "x-nestjs_zod-has-const": true
          },
          "amount": {
            "type": "number",
            "x-nestjs_zod-has-const": true
          },
          "operation": {
            "anyOf": [
              {
                "type": "string",
                "const": "income"
              },
              {
                "type": "string",
                "const": "expense"
              }
            ],
            "x-nestjs_zod-has-const": true,
            "x-nestjs_zod-empty-type": true
          },
          "method": {
            "type": "string",
            "x-nestjs_zod-has-const": true
          }
        },
        "required": [
          "title",
          "amount",
          "operation"
        ]
      },
      "EditTransactionBodyDTO": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "method": {
            "type": "string"
          }
        }
      }
    }
  }
}